[FIX] (finish incantation: add free to the incantation str and split file incantation into 2 files
