+++ src/Server/include/command.h
@@ -41,5 +41,7 @@ void fork_c(server_t *server, client_t *client, char *buffer);
 void handle_player_death(server_t *server, client_t *client);
 void check_player_starvation(server_t *server, client_t *client);
 void start_incantation(server_t *server, client_t *client, char *buffer);
+void finish_incantation(server_t *server, client_t *client);
+bool can_start_incantation(server_t *server, client_t *client);
 
 #endif /* !COMMAND_H_ */
